Refactor store interface
