+++ backend/types/concerta_store_interface.go
@@ -1,0 +1,20 @@
+package types
+
+type Store interface {
+	GetUserByEmail(email string) (*User, error)
+	GetUserByID(id uint) (*User, error)
+	CreateUser(User) error
+	GetArtistByMBID(artist string) (*Artist, error)
+	GetArtistByName(name string) (*Artist, error)
+	CreateArtist(Artist) error
+	CreateArtistIfMissing(Artist) *Artist
+	CreateVenue(Venue) error
+	CreateVenueIfMissing(Venue) *Venue
+	GetVenueByName(string) (*Venue, error)
+	CreateTour(Tour) error
+	CreateTourIfMissing(Tour) *Tour
+	GetTourByName(string) (*Tour, error)
+	CreateConcertIfMissing(Concert) *Concert
+	CreateSongIfMissing(Song) *Song
+	CreateConcertSongIfMissing(ConcertSong) *ConcertSong
+}
